Version 2.0.4 BETA
Added some Drawing is done differently, so one can see the guards circles on top instead of under another placed guards lines. Because of this, I Removed CIRCLE_SIZE and DISPLACEMENT from Guard Class and put it inside the Data Model Class since Art Gallery has to know the size of those guards. because I wanted to change how the drawing occurred Also now draws a number next to the guard showing the level of vision that guard holds.